flapping.cephfs: Add specific smb2.session tests as knownfail
Many of the tests from smb2.session torture suite are already
failing on XFS for which flapping.xfs already contains an entry
to skip any such failures. With recent update to selftest entries
from upstream, compared to flapping.xfs we are left with only one
failure(smb2.session.reauth4) which needs to be investigated before
adding it to the flapping list.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>