testhelper/smbclient: set different client guids
The smbclient high level calls make use of a global configuration which
ends up reusing the client guid. This doesn't work for us in our multi
client testing.

Make use of some low level calls at the point of setting up the
connection so that we use different client guids for each connection.

This particular workaround requires setting the connection cache
bypassing the methods available in the smbclient module. Setting this as
a separate commit to highlight the part which may break if the
underlying implementation is changed in the future.

Signed-off-by: Sachin Prabhu <sp@spui.uk>