flapping.cephfs: Add smb2.session.reauth4 as knownfail
With recent changes to smb.conf[1], smb2.session.reauth4 is expected to
fail on CephFS kclient method. Refer [1] for more details.

Also remember that this is not a problem with CephFS vfs approach but
our clone of selftest mechanism currently lacks the ability to consider
flapping lists based on variants for a particular backend.

[1] https://github.com/samba-in-kubernetes/sit-environment/pull/109

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>