containers/ltp: Update the custom list of tests
Recently ltp container image was rebuilt based on the latest sources.
Accordingly update the list of tests as per the expectation based on
the list of passing tests (from QE) on EL9 kernel versions. Few tests
had to be excluded as their corresponding updated sources requires
support for mknod, mkfifo, symlink etc. which are not available with
the way we perform the kernel CIFS mount.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>